+++ absen-ocb-be-main/src/index.js
@@ -17,6 +17,7 @@ const laporanRoute = require('./routes/laporan')
 const kpiSalaryRoute = require('./routes/kpiSalary')
 const rawAbsensiRoute = require('./routes/rawAbsensi')
 const migrationRouter = require('./routes/migration')
+const jadwalHarianRoute = require('./routes/jadwalHarian')
 // const midllewareLogRequest = require('./middleware/logs');
 const createError = require('http-errors');
 const cors = require('cors');
@@ -48,6 +49,7 @@ app.use('/api/file', laporanRoute)
 app.use('/api/management', kpiSalaryRoute)
 app.use('/api/raw-absensi', rawAbsensiRoute)
 app.use('/api/migration', migrationRouter)
+app.use('/api/jadwal-harian', jadwalHarianRoute)
 app.use(async (req, res, next) => {
     next(createError.NotFound())
 })