+++ absen-ocb-fe-main/src/App.jsx
@@ -21,6 +21,7 @@ import Profile from "./Pages/Profile";
 import AbsenKaryawan from "./Pages/AbsenKaryawan";
 import SalaryKpi from "./Pages/SalaryKpi";
 import RekapKalender from "./Pages/RekapKalender";
+import JadwalHarian from "./Pages/JadwalHarian";
 function App() {
   const [isSidebarOpen, setIsSidebarOpen] = useState(false);
   const toggleSidebar = () => {
@@ -62,6 +63,7 @@ function App() {
                     <Route path="/laporan" element={<Laporan />} />
                     <Route path="/profile" element={<Profile />} />
                     <Route path="/rekap-kalender" element={<RekapKalender />} />
+                    <Route path="/jadwal-harian" element={<JadwalHarian />} />
                   </Routes>
                 </div>
               </div>